requirements.txt
### dependency
tqdm
dvc
pandas
numpy
PyYAML
mkdocs-material
Scipy
Scikit-learn
lxml
### local packages -
-e . # << uncomment this to install from setup.py   